pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "joaolucasalves010",email = "ylucasslalves@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi[all] (>=0.135.3,<0.136.0)",
    "sqlmodel (>=0.0.38,<0.0.39)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
